dirty state is now a byte instead of a full 32 bit int

+++ src/core/MonsterController.hpp
@@ -83,7 +83,7 @@ public:
     virtual ~MonsterNetworkController() {}
     
     virtual void read(InputMemoryBitStream& imbs);
-    virtual uint16_t write(OutputMemoryBitStream& ombs, uint16_t dirtyState);
+    virtual uint8_t write(OutputMemoryBitStream& ombs, uint8_t dirtyState);
     virtual void recallCache();
-    virtual uint16_t refreshDirtyState();
+    virtual uint8_t refreshDirtyState();
 };
