+++ src/core/PlayerController.hpp
@@ -125,7 +125,7 @@ public:
     virtual ~PlayerNetworkController() {}
     
     virtual void read(InputMemoryBitStream& imbs);
-    virtual uint16_t write(OutputMemoryBitStream& ombs, uint16_t dirtyState);
+    virtual uint8_t write(OutputMemoryBitStream& ombs, uint8_t dirtyState);
     virtual void recallCache();
-    virtual uint16_t refreshDirtyState();
+    virtual uint8_t refreshDirtyState();
 };
